requirements.txt
dash
yfinance
plotly
openai
python-dotenv